requirements.txt
Flask 
Flask-SQLAlchemy 
Flask-CORS 
Flask-Migrate 
azure-storage-blob dotenv 
gunicorn